Se avanzo con las armas
se creo un gun manager
se creo una pool para las balas
las balas deberan crear su propio movimiento

+++ Assets/Guns/AttachMents/AttachMent.cs
@@ -5,7 +5,7 @@ using static GunStats;
 
 public struct AttachmentStats
 {
-    public Dictionary<StatNames, float> _Stats;
+    public Dictionary<StatNames, int> _Stats;
 }
 public abstract class Attachment : MonoBehaviour
 {
